requirements.txt
requests==2.31.0
environs==10.0.0
redis==3.2.1
python-telegram-bot==13.15
validators==0.22.0
